requirements.txt
alembic==1.13.2
annotated-types==0.6.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.6
django-bootstrap5==24.2
django-cors-headers==4.4.0
django-crispy-forms==2.1
django-extensions==3.2.3
django-filter==24.3
django-phonenumber-field==7.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dlib==19.24.4
face-recognition==1.3.0
face_recognition_models==0.3.0
Faker==26.2.0
filelock==3.15.4
fsspec==2024.6.1
geographiclib==2.0
geopy==2.4.1
huggingface-hub==0.23.5
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.5
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
phonenumbers==8.13.36
pillow==10.4.0
psycopg==3.2.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.11.0
urllib3==2.2.2
